refactor: BaseNode로 통합 및 focus 상태 관리 수정
